forms - turning one on turns the other off
I've improved the toggle a little more and added some comments.

So now the toggle works as, if you click a button the corresponding form will show, if you click that same button again it will hide. If a form is open and you click the button for the other form it will hide the current form and show the other.