functional w/ issues
currently if you use any sort of auto complete/fill then the onChange event in the text fields wont fire. Also when you press login, it will fail the first time (setting the state must be out of time or something like that) so you have to press it twice. but it will take you to the student login and this should be easily expandable on the back-end to allow for more types of users: admin, lunch shopper/maker.